Biz logic is managed through model, and the new mechanism is used to manage the online state of users.
1. 自注册和自动初始化的model管理框架
2. chat与photo、contact等model解耦
3. rpc_server业务逻辑通过自注册的model实现
4. 使用可替换的简化的用户在线状态机制
5. ...

+++ biz/dal/dao/mysql_dao/user_contacts_dao.go
@@ -205,7 +205,7 @@ func (dao *UserContactsDAO) SelectBlockedList(owner_user_id int32, limit int32) 
 
 // update user_contacts set contact_first_name = :contact_first_name, contact_last_name = :contact_last_name, is_deleted = 0 where id = :id
 // TODO(@benqi): sqlmap
-func (dao *UserContactsDAO) UpdateContactNameById(contact_first_name string, contact_last_name string, id int32) int64 {
+func (dao *UserContactsDAO) UpdateContactNameById(contact_first_name string, contact_last_name string, id int64) int64 {
 	var query = "update user_contacts set contact_first_name = ?, contact_last_name = ?, is_deleted = 0 where id = ?"
 	r, err := dao.db.Exec(query, contact_first_name, contact_last_name, id)
